% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{calculateLocal}
\alias{calculateLocal}
\title{Calculate local Moran's I or local Geary's C}
\usage{
calculateLocal(response, neighboursList, weightStyle = c("B", "W"),
  localMethod = c("moran", "geary"))
}
\arguments{
\item{response}{numeric vector; response of interest for analysis}

\item{neighboursList}{object as returned by poly2nb()}

\item{weightStyle}{character, indicating style for assigning neighbour weights;
one of \code{c("B", "W")} respectively row-standardized (sum of weights per 
area is 1) or binary (neighbour is 1, non-neighbour is 0); default is "B"}

\item{localMethod}{method for local values calculation; one of 
\code{c("moran", "geary")}; default is "moran"}
}
\value{
dataframe with 'statistics' ie local moran/geary statistic, 
regions with missing response or without neighbours will have value NA;
and 'pValues' for localMethod moran only ie p-value of local moran statistic
}
\description{
Calculate local Moran's I or local Geary's C
}

