% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{collapseData}
\alias{collapseData}
\title{Collapse data according to the user-defined variable}
\usage{
collapseData(inputData, collapseVariable, responsePositive = NULL,
  responseTotal = NA, covariates = NULL, functionCovariates = NULL)
}
\arguments{
\item{inputData}{data frame, the original data}

\item{collapseVariable}{character vector, variable names in inputData with 
respect to which the data should be collapsed}

\item{responsePositive}{character, variable name in inputData that indicates
the number of positive responses; default value is NULL}

\item{responseTotal}{character, variable name in inputData that indicates 
the total number of observations for the response; default value is NA}

\item{covariates}{character vector, variable names in inputData of the 
covariates that should be summarized when collapsing the data}

\item{functionCovariates}{character vector, the function for summarizing each
of the covariates; each one of \code{c("mean", "max", "min", "modus", 
"relativeFrequency")}; default is NULL}

\item{makeResponses}{boolean, whether to create response variables or not;
default value is TRUE}
}
\value{
data frame, with the unique rows of the inputData (ie collapsed
inputData excluding the response), (numberPositive) the number of positive 
response values and (numberObservations) the total number of response values
}
\description{
Collapse data according to the user-defined variable
}

