% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{createClasses}
\alias{createClasses}
\title{Create classes for response according to breaksStyle}
\usage{
createClasses(response, breaksStyle = c("quantile", "jenks", "fixed",
  "quadrant"), nClasses, fixedBreaks, intervalClosure = c("left", "right"),
  localValues, significance = NA)
}
\arguments{
\item{response}{numeric vector; response of interest for analysis}

\item{breaksStyle}{character, defines the style for creating classes; one of 
\code{c("quantile", "jenks", "fixed", "quadrant")}; default is quantile;
quadrants are for calculated local values, for the others see classIntervals()}

\item{nClasses}{numeric; number of classes that should be created;
value is ignored if breaksStyle is "quadrants"}

\item{fixedBreaks}{numeric vector with breaks values; ignored unless 
breaksStyle is "fixed"}

\item{intervalClosure}{character; one of \code{c("left", "right")}; 
default is left}

\item{localValues}{object as returned by calculateLocal()}

\item{significance}{numeric value, used to identify significant local values;
default is NA; ignored unless breaksStyle is "quadrants"}
}
\value{
vector of factor levels with same length as response
}
\description{
Create classes for response according to breaksStyle
}

