% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{makeDates}
\alias{makeDates}
\title{Create dates for given year, month and day}
\usage{
makeDates(inputData, timeNames)
}
\arguments{
\item{inputData}{data frame, the original data}

\item{timeNames}{character vector of max length 3, variable names in inputData 
for year, month and day (in that order); month or day may be missing and are
then replaced by value 1}
}
\value{
dates vector, the date based on given year, month and day in inputData
}
\description{
Create dates for given year, month and day
}

