% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{makeFormulaInla}
\alias{makeFormulaInla}
\title{Construct the formula for the inla() model fit}
\usage{
makeFormulaInla(responseName, covariateNames = NULL, timeEffect = c("none",
  "saturated", "linear", "rw1", "ar1", "rw2", "all"),
  interactionEffect = c("none", "type1", "type2", "type3", "type4"))
}
\arguments{
\item{responseName}{character string, name of the response to analyse}

\item{covariateNames}{character vector, names of the covariates for the 
inla model; default is NULL}

\item{timeEffect}{character string, defines the time effect; one of 
\code{c("none", "saturated", "linear", "rw1", "ar1", "rw2", "all")};
default is "none"}

\item{interactionEffect}{character string, defines the interaction effect;
one of c("none", "type1", "type2", "type3", "type4"); default is "none"}
}
\value{
character string, with the formula for inla(); if timeEffect is "all"
a vector of character strings for all possible models is returned
}
\description{
Construct the formula for the inla() model fit
}

