% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{makeLagDates}
\alias{makeLagDates}
\title{Add lag time to dates}
\usage{
makeLagDates(dates, lagTime, timeUnit = c("Year", "Month", "Day"))
}
\arguments{
\item{dates}{dates vector, input values for the dates}

\item{lagTime}{integer, the number of years/months/days that should
be added to the dates}

\item{timeUnit}{character, the timeUnit for addition; should be one of
\code{c("Year", "Month", "Day")}; default value is "Year"}
}
\value{
dates vector, the lagged dates ie original dates with added number 
of years/months/days
}
\description{
Add lag time to dates
}

