% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{makeSpaceTimeData}
\alias{makeSpaceTimeData}
\title{Collapse and extend data frame to have one row for each time unit at the 
spatial level of interest}
\usage{
makeSpaceTimeData(mergedData, timeNames, timeUnit = c("Year", "Month", "Week",
  "Day"), spatialUnit, inlaVariables = TRUE, responsePositive,
  responseTotal = NA, covariates = NULL, functionCovariates = NULL)
}
\arguments{
\item{timeNames}{character vector, names of the time variables for 
year, month and day (in that order)}

\item{timeUnit}{character string that determines the time level of interest; 
should be one of \code{c("Year", "Month", "Week", "Day")}; default is "year"}

\item{spatialUnit}{character string, variable name in the infoData that 
determines the spatial level of interest}

\item{inlaVariables}{boolean, if TRUE extra variables at space and time level
are created, as needed for inla(); default value is TRUE}

\item{responsePositive}{character, variable name in inputData that indicates
the number of positive responses}

\item{responseTotal}{character, variable name in inputData that indicates 
the total number of observations for the response; default value is NA}

\item{covariates}{character vector, variable names in inputData of the 
covariates that should be summarized when collapsing the data; default is NULL}

\item{functionCovariates}{character vector, the function for summarizing each
of the covariates; each one of \code{c("mean", "max", "min", "modus")}; 
if one function is specified for a vector of covariates that same function 
is used for all covariates; default is NULL}

\item{infoData}{data frame at the spatial level of interest, contains 
the timeNames (for year, month and day) and spatialUnit (for NUTS0, NUTS1, 
NUTS2, NUTS3, LAU1 or LAU2)}
}
\value{
extended data frame of infoData with variables: spatialUnit and 
timeUnit with unique value for resp. each spatial unit and time unit; 
if inlaVariables is TRUE also spatialTimeUnit which is combination of both
}
\description{
Collapse and extend data frame to have one row for each time unit at the 
spatial level of interest
}

