% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{mergeAllData}
\alias{mergeAllData}
\title{Merge the temporal, spatial and covarites data at lowest space and time level}
\usage{
mergeAllData(temporalData, spatialData, covariatesData, responseVariable,
  timeResolution = c("Year", "Month", "Day", "Date"), dateFormat,
  spaceResolution = c("NUTS0", "NUTS1", "NUTS2", "NUTS3", "LAU1", "LAU2"),
  timeNamesTemporal, spatialNameTemporal, covariatesTemporal = NULL,
  functionCovariatesTemporal = NULL, timeNamesExtra, spatialNameExtra,
  covariatesExtra = NULL, functionCovariatesExtra = NULL, lagTime = 0,
  covariatesSpatial = NULL)
}
\arguments{
\item{temporalData}{data frame, the temporal data}

\item{spatialData}{data frame, the spatial data}

\item{covariatesData}{data frame, the covariates data}

\item{responseVariable}{character, name of the binary response in the temporal data}

\item{timeResolution}{character, indicates the lowest time level;
one of \code{c("Year", "Month", "Day", "Date")}}

\item{dateFormat}{character, indicates the format of dates; is only used for
timeResolution "Date"}

\item{spaceResolution}{character, indicates the lowest space level;
one of \code{c("NUTS0", "NUTS1", "NUTS2", "NUTS3", "LAU1", "LAU2")}}

\item{timeNamesTemporal}{character vector, the variable names in temporalData
for the time variables}

\item{spatialNameTemporal}{character, the variable name in temporalData for
the space level}

\item{covariatesTemporal}{character vector, the variable names in temporalData
for the covariates}

\item{functionCovariatesTemporal}{character vector, the names of the functions
used to summarize the covariates in temporalData}

\item{timeNamesExtra}{character vector, the variable names in covariatesData
for the time variables}

\item{spatialNameExtra}{character vector, the variable name in covariatesData
for the space level}

\item{covariatesExtra}{character vector, the variable names in covariatesData
for the covariates}

\item{functionCovariatesExtra}{character vector, the names of the functions
used to summarize the covariates in covariatesData}

\item{lagTime}{integer, the number of years/months/days that should
be added to the dates; specified in units of timeResolution}

\item{covariatesSpatial}{character vector, the variable names in spatialData
for the covariates}
}
\value{
data frame, merged temporal, spatial and covariates data; variable names
are the spatial and time levels from highest to the resolution level, 
response variables and covariates from each of the data frames to be retained
}
\description{
Merge the temporal, spatial and covarites data at lowest space and time level
}

