% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{modus}
\alias{modus}
\title{Calculate the modus}
\usage{
modus(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector for which the modus should be calculated}

\item{na.rm}{boolean, whether missing values can be ignored; default is TRUE}
}
\value{
the modus of x
}
\description{
Calculate the modus
}

