% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{performInla}
\alias{performInla}
\title{Fit space-time model with inla()}
\usage{
performInla(spaceTimeData, formulaInla, neighboursFileName, weighted = FALSE,
  verbose = FALSE)
}
\arguments{
\item{spaceTimeData}{data frame with all necessary data for the inla() model fitting;
as returned from function makeSpaceTimeData()}

\item{formulaInla}{character string, with formula returned by formulaInla()
and used as first argument of inla()}

\item{neighboursFileName}{character string indicates the file name (without 
extension) for the saved INLA neighbours list}

\item{weighted}{boolean, whether weighted inla() regression should be 
performed or not; default is FALSE}

\item{verbose}{boolean, whether to print info during the inla analysis;
default value is FALSE}
}
\value{
the output as generated by the inla() procedure; if the length of 
formulaInla is larger than 1, only the model with best time-effect will 
be returned as well as all models' dic, logScore and convergence status
}
\description{
Fit space-time model with inla()
}

