% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{performKriging}
\alias{performKriging}
\title{Perform ordinary kriging: region center used for modelling}
\usage{
performKriging(response, shapeData, variogramModel = NA)
}
\arguments{
\item{response}{numeric vector, response in the formula for defining kriging}

\item{shapeData}{a spatialPolygonsDataFrame object as returned from 
readShapeData() function}

\item{variogramModel}{object of class variogramModel as returned by function
vgm(); if missing autoKrige() is performed, default value is NA}
}
\value{
list of gstatVariogram, variogramModel and SpatialPolygonsDataFrame 
objects; the first is an extended data.frame with the sample variogram as 
returned by variogram(); the second is an extended data.frame with model 
information on the variogram, as returned from vgm(); the latter contains a 
data frame with for each region the predicted values 'var1.pred'; their 
variance 'var1.var' and standard deviation 'var1.stdev'
}
\description{
Perform ordinary kriging: region center used for modelling
}

