% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{readDates}
\alias{readDates}
\title{Read dates to generat year, month and day}
\usage{
readDates(dates, dateFormat, timeUnit = c("Year", "Month", "Day"))
}
\arguments{
\item{dates}{Date vector, the date from which year, month and day will 
be derived}

\item{dateFormat}{character, indicates the format of dates}

\item{timeUnit}{character, indicates to which time unit the dates should be
translated}
}
\value{
data frame with Year, Month and Day (or less depending on timeUnit) 
and the original Date
}
\description{
Read dates to generat year, month and day
}

