% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{relativeFrequency}
\alias{relativeFrequency}
\title{Calculate the relative frequency}
\usage{
relativeFrequency(x)
}
\arguments{
\item{x}{factor vector for which the relative frequency of each of its values should
be calculated}
}
\value{
the relative frequencies of x
}
\description{
Calculate the relative frequency
}

