% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{saveNeighboursInla}
\alias{saveNeighboursInla}
\title{Save neighboursList for use with inla()}
\usage{
saveNeighboursInla(neighboursList, weightStyle = c("B", "W"),
  fileName = "adjacencyINLA")
}
\arguments{
\item{neighboursList}{object of class "nb", a neighbours list}

\item{weightStyle}{character, indicating style for assigning neighbour weights;
one of \code{c("B", "W")} respectively row-standardized (sum of weights per 
area is 1) or binary (neighbour is 1, non-neighbour is 0); default is "B"}

\item{fileName}{character string, specifies the file name (without extension)
to save neighbours info required by inla(); default is "adjacencyINLA"}
}
\value{
NULL. Writes file with neighbours as required by inla() for the 
spatial models in temporary directory
}
\description{
Save neighboursList for use with inla()
}

