% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{selectPolygons}
\alias{selectPolygons}
\title{Select subset of polygons for the given spatial level}
\usage{
selectPolygons(allPolygons, spatialLevel = c("NUTS0", "NUTS1", "NUTS2",
  "NUTS3", "LAU1", "LAU2"), doSelect = NULL)
}
\arguments{
\item{allPolygons}{SpatialPolygonsDataFrame object as returned from 
makeUnionPolygons(); contains polygons at different spatial levels}

\item{spatialLevel}{character string, indicates the spatial level at which
all polygons should be selected; one of c("NUTS0", "NUTS1", "NUTS2", "NUTS3", 
"LAU1", "LAU2"); ignored if doSelect is not NULL; default value is "NUTS0"}

\item{doSelect}{boolean vector with length the number of polygons in allPolygons, 
indicates whether to select the polygon; if NULL, spatial level should be 
defined; default value is NULL}
}
\value{
SpatialPolygonsDataFrame object, which is subset of allPolygons 
containing only polygons at the specified spatial level
}
\description{
Select subset of polygons for the given spatial level
}

